+++ res/resource.h
@@ -229,6 +229,7 @@
 #define IDC_PDF_FORM_EDIT        1172
 #define IDC_PDF_FORM_SAVE        1173
 #define IDM_FILE_FILL_FORM       7420
+#define IDM_FILE_SIGN_PDF        7421
 #define IDD_COMMENTS             1160
 #define IDC_COMMENTS_LIST        1161
 #define IDC_COMMENTS_DELETE      1162